#39 added counters to goal/view and made an Add method for adding from this interface

+++ migrations/m161222_072411_counters.php
@@ -21,7 +21,7 @@ class m161222_072411_counters extends Migration
             'id' => $this->primaryKey(),
             'counter_id' => $this->integer()->notNull(),
             'time' => $this->dateTime()->notNull(),
-            'value' => $this->integer()->notNull(),
+            'value' => $this->float()->notNull(),
             'description' => $this->text()->notNull(),
         ], 'CHARACTER SET utf8 COLLATE utf8_unicode_ci ENGINE=InnoDB');
